requirements.txt
imageio==2.25.1
imageio-ffmpeg==0.4.8
natsort==8.2.0
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
psutil==5.9.4
